+++ IqSoft.CP.AdminWebApiCore/Hubs/BaseHub.cs
@@ -305,12 +305,12 @@ namespace IqSoft.CP.AdminWebApi.Hubs
 				var partnerAccess = partnerBll.GetPermissionsToObject(new CheckPermissionInput
 				{
 					Permission = Constants.Permissions.ViewPartner,
-					ObjectTypeId = ObjectTypes.Partner
+					ObjectTypeId = (int)ObjectTypes.Partner
 				});
 				var clientAccess = partnerBll.GetPermissionsToObject(new CheckPermissionInput
 				{
 					Permission = Constants.Permissions.ViewClient,
-					ObjectTypeId = ObjectTypes.Client
+					ObjectTypeId = (int)ObjectTypes.Client
 				});
 				if ((!partnerAccess.HaveAccessForAllObjects && partnerAccess.AccessibleObjects.All(x => x != result.PartnerId) ||
 					(!clientAccess.HaveAccessForAllObjects && clientAccess.AccessibleObjects.All(x => x != result.ClientId))))
