+++ IqSoft.CP.BLLCore/Services/BetShopBll.cs
@@ -177,7 +177,7 @@ namespace IqSoft.CP.BLL.Services
             CheckPermissionToSaveObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.CreateBetShopGroup,
-                ObjectTypeId = ObjectTypes.BetShopGroup,
+                ObjectTypeId = (int)ObjectTypes.BetShopGroup,
                 ObjectId = betShopGroup.Id
             });
             var currentTime = GetServerDate();
@@ -316,7 +316,7 @@ namespace IqSoft.CP.BLL.Services
             CheckPermissionToSaveObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.CreateBetShop,
-                ObjectTypeId = ObjectTypes.BetShop,
+                ObjectTypeId = (int)ObjectTypes.BetShop,
                 ObjectId = betShop.Id
             });
             var currentTime = GetServerDate();
@@ -375,7 +375,7 @@ namespace IqSoft.CP.BLL.Services
             CheckPermissionToSaveObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.CreateCashDesk,
-                ObjectTypeId = ObjectTypes.CashDesk,
+                ObjectTypeId = (int)ObjectTypes.CashDesk,
                 ObjectId = cashDesk.Id
             });
 
@@ -421,7 +421,7 @@ namespace IqSoft.CP.BLL.Services
             CheckPermissionToSaveObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.DeleteBetShopGroup,
-                ObjectTypeId = ObjectTypes.BetShopGroup,
+                ObjectTypeId = (int)ObjectTypes.BetShopGroup,
                 ObjectId = id
             });
 
@@ -458,7 +458,7 @@ namespace IqSoft.CP.BLL.Services
                 var checkPermissionResult = GetPermissionsToObject(new CheckPermissionInput
                 {
                     Permission = Constants.Permissions.ViewBetShop,
-                    ObjectTypeId = ObjectTypes.BetShop,
+                    ObjectTypeId = (int)ObjectTypes.BetShop,
                     UserId = userId
                 });
                 if (!checkPermissionResult.HaveAccessForAllObjects && !checkPermissionResult.AccessibleObjects.AsEnumerable().Contains(id))
@@ -476,12 +476,12 @@ namespace IqSoft.CP.BLL.Services
                 var betshoppermishion = GetPermissionsToObject(new CheckPermissionInput
                 {
                     Permission = Constants.Permissions.ViewBetShop,
-                    ObjectTypeId = ObjectTypes.BetShop
+                    ObjectTypeId = (int)ObjectTypes.BetShop
                 });
                 var partnerAccess = GetPermissionsToObject(new CheckPermissionInput
                 {
                     Permission = Constants.Permissions.ViewPartner,
-                    ObjectTypeId = ObjectTypes.Partner
+                    ObjectTypeId = (int)ObjectTypes.Partner
                 });
 
                 filter.CheckPermissionResuts.Add(new CheckPermissionOutput<BetShop>
@@ -507,12 +507,12 @@ namespace IqSoft.CP.BLL.Services
             var betshoppermishion = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewBetShop,
-                ObjectTypeId = ObjectTypes.BetShop
+                ObjectTypeId = (int)ObjectTypes.BetShop
             });
             var partnerAccess = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewPartner,
-                ObjectTypeId = ObjectTypes.Partner
+                ObjectTypeId = (int)ObjectTypes.Partner
             });
 
             filter.CheckPermissionResuts.Add(new CheckPermissionOutput<fnBetShops>
@@ -535,7 +535,7 @@ namespace IqSoft.CP.BLL.Services
             var checkP = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewCashDesk,
-                ObjectTypeId = ObjectTypes.CashDesk
+                ObjectTypeId = (int)ObjectTypes.CashDesk
             });
 
             filter.CheckPermissionResuts = new List<CheckPermissionOutput<CashDesk>>
@@ -554,7 +554,7 @@ namespace IqSoft.CP.BLL.Services
             var checkP = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewCashDesk,
-                ObjectTypeId = ObjectTypes.CashDesk
+                ObjectTypeId = (int)ObjectTypes.CashDesk
             });
 
             filter.CheckPermissionResuts = new List<CheckPermissionOutput<fnCashDesks>>
@@ -573,13 +573,13 @@ namespace IqSoft.CP.BLL.Services
             var checkP = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewBetShop,
-                ObjectTypeId = ObjectTypes.BetShopGroup
+                ObjectTypeId = (int)ObjectTypes.BetShopGroup
             });
 
             var partnerAccess = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewPartner,
-                ObjectTypeId = ObjectTypes.Partner
+                ObjectTypeId = (int)ObjectTypes.Partner
             });
 
             filter.CheckPermissionResuts = new List<CheckPermissionOutput<BetShopGroup>>
@@ -771,7 +771,7 @@ namespace IqSoft.CP.BLL.Services
             var partnerAccess = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewPartner,
-                ObjectTypeId = ObjectTypes.Partner
+                ObjectTypeId = (int)ObjectTypes.Partner
             });
             if (!partnerAccess.HaveAccessForAllObjects && partnerAccess.AccessibleObjects.All(x => x != betShop.PartnerId))
                 throw CreateException(LanguageId, Constants.Errors.DontHavePermission);
@@ -779,7 +779,7 @@ namespace IqSoft.CP.BLL.Services
             var betShopPermission = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewBetShop,
-                ObjectTypeId = ObjectTypes.BetShop
+                ObjectTypeId = (int)ObjectTypes.BetShop
             });
 
             if (!betShopPermission.HaveAccessForAllObjects && betShopPermission.AccessibleObjects.All(x => x != cashDesk.BetShopId))
@@ -929,7 +929,7 @@ namespace IqSoft.CP.BLL.Services
             CheckPermissionToSaveObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ChangeBetShopLimit,
-                ObjectTypeId = ObjectTypes.BetShop,
+                ObjectTypeId = (int)ObjectTypes.BetShop,
                 ObjectId = betShop.Id
             });
 
@@ -973,13 +973,13 @@ namespace IqSoft.CP.BLL.Services
             var betshoppermishion = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewBetShop,
-                ObjectTypeId = ObjectTypes.BetShop
+                ObjectTypeId = (int)ObjectTypes.BetShop
             });
 
             var partnerAccess = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewPartner,
-                ObjectTypeId = ObjectTypes.Partner
+                ObjectTypeId = (int)ObjectTypes.Partner
             });
 
             filter.CheckPermissionResuts.Add(new CheckPermissionOutput<fnAdminShiftReport>
@@ -1057,13 +1057,13 @@ namespace IqSoft.CP.BLL.Services
             var betshoppermishion = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewBetShop,
-                ObjectTypeId = ObjectTypes.BetShop
+                ObjectTypeId = (int)ObjectTypes.BetShop
             });
 
             var partnerAccess = GetPermissionsToObject(new CheckPermissionInput
             {
                 Permission = Constants.Permissions.ViewPartner,
-                ObjectTypeId = ObjectTypes.Partner
+                ObjectTypeId = (int)ObjectTypes.Partner
             });
 
             filter.CheckPermissionResuts.Add(new CheckPermissionOutput<fnAdminShiftReport>
