+++ IqSoft.CP.BLLCore/Services/LanguageBll.cs
@@ -53,7 +53,7 @@ namespace IqSoft.CP.BLL.Services
                 var checkPermissionResult = GetPermissionsToObject(new CheckPermissionInput
                 {
                     Permission = Constants.Permissions.CreatePartnerLanguageSetting,
-                    ObjectTypeId = ObjectTypes.Language,
+                    ObjectTypeId = (int)ObjectTypes.Language,
                     ObjectId = partnerLanguageSetting.Id
                 });
